Move logic for assembling page template data into its own function, so it can be more easily tested
