+++ lib/panini.js
@@ -8,6 +8,7 @@ const globWatcher = require('glob-watcher');
 const ora = require('ora');
 const loadEngine = require('./load-engine');
 const render = require('./render');
+const processRoot = require('./process-root');
 
 module.exports = class Panini extends EventEmitter {
   /**
@@ -126,6 +127,55 @@ module.exports = class Panini extends EventEmitter {
     this.on('error', () => {
       this.spinner.fail('There was an error parsing a page.');
     });
+  }
+
+  /**
+   * Assemble the template data for a page. Page data is prioritized in the following order, from
+   * lowest to highest:
+   *   - Loaded data files
+   *   - Vinyl file attributes
+   *   - Page Front Matter
+   *   - Page constants:
+   *     - `page`: basename of the page
+   *     - `layout`: dervied layout of the page
+   *     - `root`: path prefix to root directory from this page
+   *
+   * @param {Object} attributes - Page Front Matter.
+   * @param {Object} file - Vinyl file containing the page.
+   * @returns {Object} Page template data.
+   */
+  getPageData(attributes, file) {
+    // Determine which layout to use
+    let layout;
+    if (this.engine.supports('layouts')) {
+      const basePath = path.relative(
+        path.join(this.options.input, this.options.pages),
+        path.dirname(file.path)
+      );
+      layout =
+        attributes.layout ||
+        this.options.pageLayouts[basePath] ||
+        'default';
+    }
+
+    return assign(
+      {},
+      // Global data
+      this.engine.data,
+      // Data from Gulp stream plugins
+      file.data || {},
+      // Page-specific data
+      attributes,
+      // Constants
+      {
+        // Basename of file
+        page: path.basename(file.path, '.html'),
+        // Layout used by this page
+        layout,
+        // Path prefix to root directory
+        root: processRoot(file.path, path.join(this.options.input, this.options.pages))
+      }
+    );
   }
 
   render() {
