+++ lib/render.js
@@ -1,10 +1,7 @@
 'use strict';
 
-const path = require('path');
-const assign = require('lodash.assign');
 const fm = require('front-matter');
 const through = require('through2');
-const processRoot = require('./process-root');
 
 /**
  * Create a transform stream to render a set of pages to HTML.
@@ -39,40 +36,10 @@ function parse(file, enc, cb) {
   // Get the HTML for the current page and layout
   const page = fm(file.contents.toString());
 
-  // Determine which layout to use
-  let layout;
+  // Assemble data for template
+  const pageData = this.getPageData(page.attributes, file);
 
-  if (this.engine.supports('layouts')) {
-    const basePath = path.relative(
-      path.join(this.options.input, this.options.pages),
-      path.dirname(file.path)
-    );
-    layout =
-      page.attributes.layout ||
-      this.options.pageLayouts[basePath] ||
-      'default';
-  }
-
-  // Build Handlebars context
-  const pageData = assign(
-    {},
-    // Global data
-    this.engine.data,
-    // Data from Gulp stream plugins
-    file.data || {},
-    // Page-specific data
-    page.attributes,
-    // Constants
-    {
-      // Basename of file
-      page: path.basename(file.path, '.html'),
-      // Layout used by this page
-      layout,
-      // Path prefix to root directory
-      root: processRoot(file.path, path.join(this.options.input, this.options.pages))
-    }
-  );
-
+  // Store page for later rendering
   this.pages.push([file, page.body, pageData]);
 
   cb();
